pyproject.toml
[tool.poetry]
name = "pisicap"
version = "0.1.0"
description = "SICAP în Piton"
authors = ["tsundokul <timofte17daniel@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.26.0"
python-dateutil = "^2.8.2"
tenacity = "^8.2.3"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
